requirements.txt
tensorflow>=1.1.0
grpcio==1.7.0
hydro-serving-grpc==0.0.7
docker==2.7.0